베이스 거의 완성...
이제 남은 건 입력 핸들러 스크립트에 BT에 ML_Agent 적용에 HUD까지...

+++ Assets/scripts/hp_bar_control.cs
@@ -11,6 +11,6 @@ public class HealthBarcontrol : MonoBehaviour
     void Update()
     {
         int currentHealth = amI1p ? gameManager.p1HP : gameManager.p2HP;
-        fillImage.fillAmount = (float)currentHealth / (float)maxHealth;
+        fillImage.fillAmount = (float)currentHealth / maxHealth;
     }
 }
